// client/src/types/offer.ts
export type OfferLocation = {
    latitude: number;
    longitude: number;
    zoom: number;
};

export type CityOffer = {
    name: string;
    location: OfferLocation;
};

export type HostOffer = {
    name: string;
    avatarUrl: string;
    isPro: boolean;
}

export type FullOffer = {
    id: string;
    title: string;
    type: string;
    price: number;
    city: CityOffer;
    location: OfferLocation;
    isFavorite: boolean;
    isPremium: boolean;
    rating: number;
    description: string;
    rooms: number;
    goods: string[];
    host: HostOffer;
    images: string[];
    maxAdults: number;
}

export type OffersList = {
    id: string;
    title: string;
    type: string;
    price: number;
    city: CityOffer;
    location: OfferLocation;
    isFavorite: boolean;
    isPremium: boolean;
    rating: number;
    previewImage: string;
}

export type User = {
    name: string;
    avatarUrl: string;
    isPro: boolean;
}

export type Review = {
    id: string;
    offerId: string;
    date: string;
    user: User;
    comment: string;
    rating: number;
}

export type ToggleFavoriteData = {
  offerId: string;
  status: 0 | 1; 
};